refactor(audit): streamline audit log components and improve user detail page
- Removed unused imports and functions from the audit log and user detail pages to enhance code clarity and maintainability.
- Integrated new components, `AuditChangeDetailModal` and `AuditLogSharedCells`, to improve the display of audit log details.
- Updated the layout of the user detail page for better readability and user experience, including adjustments to table headers and styles.
- Simplified the logic for displaying audit logs and user snapshots, ensuring a more efficient rendering process.
