+++ packages/admin/app/gateway/audit-logs/page.tsx
@@ -5,13 +5,11 @@
  */
 import { useTranslations } from 'next-intl';
 import { useState, useEffect, useCallback, useRef } from 'react';
-import Link from 'next/link';
 import { readApiJson } from '@/lib/api-json';
 import {
   API_KEY_BUDGET_AUDIT_ACTOR_TYPES,
   API_KEY_BUDGET_AUDIT_EVENT_TYPES,
   API_KEY_BUDGET_AUDIT_SOURCE_CHANNELS,
-  USER_AUDIT_ACTOR_KINDS,
   type GatewayApiKeyBudgetAuditLog,
 } from '@/lib/types';
 import { GatewayTimeRangePicker } from '@/components/GatewayTimeRangePicker';
@@ -22,24 +20,11 @@ import {
   type GatewayTimeRangeValue,
 } from '@/lib/analytics-range';
 import { useReplaceListPageQuery } from '@/lib/use-replace-list-query';
-import { formatGatewayDateTime } from '@/lib/datetime';
-import { formatGatewayMoneyCode, formatGatewayMoneyCodeSigned } from '@/lib/format-gateway-currency';
-import { GATEWAY_MONEY_DECIMAL_PLACES } from '@/lib/gateway-money';
-import { summarizeWalletSnapshotDiff, WALLET_AUDIT_SNAPSHOT_FIELDS } from '@/lib/audit-user-snapshot-diff';
-import { AuditWalletPlanBlock } from '@/components/AuditWalletPlanBlock';
 import { MultiSelectDropdown } from '@/components/MultiSelectDropdown';
+import { AuditChangeDetailModal, AuditLogSharedCells } from '@/components/AuditLogSharedCells';
+import { auditEnumLabel } from '@/lib/audit-log-display';
 import { useBillingCurrency } from '@/lib/use-billing-currency';
 import { useGatewayDateTime } from '@/lib/use-gateway-datetime';
-
-/** 已在 Budget / Period plan / Wallet 列展示的快照字段，不在「User change detail」重复 */
-const OMIT_AUDIT_LOG_SNAPSHOT_FIELDS = [
-	'budget_spent',
-	'budget_max',
-	'budget_base',
-	'budget_period',
-	'budget_reset_at',
-	...WALLET_AUDIT_SNAPSHOT_FIELDS,
-] as const;
 
 const DEFAULT_AUDIT_LOG_EVENT_TYPES = API_KEY_BUDGET_AUDIT_EVENT_TYPES.filter((type) => type !== 'usage_charge');
 const AUDIT_LOG_EVENT_TYPE_SET = new Set<string>(API_KEY_BUDGET_AUDIT_EVENT_TYPES);
@@ -47,21 +32,9 @@ const DEFAULT_AUDIT_LOG_ACTOR_TYPES = [...API_KEY_BUDGET_AUDIT_ACTOR_TYPES];
 const AUDIT_LOG_ACTOR_TYPE_SET = new Set<string>(API_KEY_BUDGET_AUDIT_ACTOR_TYPES);
 const DEFAULT_AUDIT_LOG_SOURCE_CHANNELS = [...API_KEY_BUDGET_AUDIT_SOURCE_CHANNELS];
 const AUDIT_LOG_SOURCE_CHANNEL_SET = new Set<string>(API_KEY_BUDGET_AUDIT_SOURCE_CHANNELS);
-/** 完整前缀目录（含历史 `admin:`），表格徽章仍能识别老 Master Key 行。 */
-const AUDIT_LOG_ACTOR_KIND_SET = new Set<string>(USER_AUDIT_ACTOR_KINDS);
 /** 筛选 UI 不含历史 Master Key；全选时不传 `actor_kind`，默认列表仍含老 `admin:` 行。 */
 const AUDIT_LOG_ACTOR_KIND_FILTERS = ['console', 'admin_key', 'system', 'service'] as const;
 const AUDIT_LOG_ACTOR_KIND_FILTER_SET = new Set<string>(AUDIT_LOG_ACTOR_KIND_FILTERS);
-
-function auditEnumLabel(
-  t: ReturnType<typeof useTranslations>,
-  group: 'eventTypes' | 'sourceChannels' | 'actorTypes' | 'actorKinds',
-  value: string | null | undefined,
-): string {
-  if (value == null || value === '') return '—';
-  const key = `${group}.${value}`;
-  return t.has(key) ? t(key) : value;
-}
 
 type AuditLogFilterOptions = {
   reasonCodes: string[];
@@ -98,16 +71,6 @@ function normalizeAuditActorKinds(values: string[]): string[] {
       if (AUDIT_LOG_ACTOR_KIND_FILTER_SET.has(value) && !normalized.includes(value)) normalized.push(value);
     });
   return normalized;
-}
-
-/** `actor_id` 形如 `<kind>:<identifier>`；未知或缺失前缀时 kind 为 null，整串按标识符展示。 */
-function parseAuditActorId(actorId: string | null | undefined): { kind: string | null; identifier: string } {
-  if (!actorId) return { kind: null, identifier: '' };
-  const separator = actorId.indexOf(':');
-  if (separator === -1) return { kind: null, identifier: actorId };
-  const kind = actorId.slice(0, separator);
-  if (!AUDIT_LOG_ACTOR_KIND_SET.has(kind)) return { kind: null, identifier: actorId };
-  return { kind, identifier: actorId.slice(separator + 1) };
 }
 
 function normalizeAuditSourceChannels(values: string[]): string[] {
@@ -163,300 +126,6 @@ function appendAuditReasonCodeParams(params: URLSearchParams, reasonCodes: strin
   const isAllSelected = allReasonCodes.length > 0 && allReasonCodes.every((reasonCode) => selectedSet.has(reasonCode));
   if (reasonCodes.length === 0 || isAllSelected) return;
   reasonCodes.forEach((reasonCode) => params.append('reason_code', reasonCode));
-}
-
-/** change_payload 展开行：去掉已由 Budget / Period plan / Time / Event 列展示的键 */
-function shouldOmitChangePayloadDisplayLine(line: string): boolean {
-	const colon = line.indexOf(':');
-	const key = (colon === -1 ? line : line.slice(0, colon)).trim();
-	if (!key) return false;
-	if (key.startsWith('before_budget_') || key.startsWith('after_budget_')) return true;
-	if (['actor_id', 'reason_code', 'reason_text', 'source', 'correlation_id'].includes(key)) return true;
-	return false;
-}
-
-function formatSignedMoney(value: number, currency: string): string {
-  return formatGatewayMoneyCodeSigned(value, currency, GATEWAY_MONEY_DECIMAL_PLACES);
-}
-
-function formatBudgetMax(value: number | null, currency: string, noLimitLabel = 'no limit'): string {
-  if (value == null) return noLimitLabel;
-  return formatGatewayMoneyCode(value, currency, GATEWAY_MONEY_DECIMAL_PLACES);
-}
-
-function formatAuditTime(iso: string | null | undefined, timeZone: string): string {
-  if (iso == null || iso === '') return '—';
-  return formatGatewayDateTime(iso, timeZone);
-}
-
-function shortId(id: string | null | undefined): string {
-  if (id == null || id === '') return '—';
-  if (id.length < 14) return id;
-  return `${id.slice(0, 8)}…${id.slice(-4)}`;
-}
-
-/** Reason 行：code / text 并存且不同时压缩为一行「code · text」，否则单行。 */
-function auditReasonOneLine(reasonCode: string | null | undefined, reasonText: string | null | undefined): {
-	line: string;
-	isMono: boolean;
-	title: string;
-} {
-	const rc = (reasonCode ?? '').trim();
-	const rt = (reasonText ?? '').trim();
-	if (!rc && !rt) return { line: '—', isMono: true, title: '' };
-	if (rc && rt && rc !== rt) {
-		const line = `${rc} · ${rt}`;
-		return { line, isMono: false, title: line };
-	}
-	const single = rt || rc;
-	return { line: single, isMono: !!rc && !rt, title: single };
-}
-
-/** 从 `change_payload` 解析的扩展字段（与 `@octafuse/core` `mergeUserAuditChangePayload` 写入结构对齐） */
-function auditDisplayExtras(item: GatewayApiKeyBudgetAuditLog) {
-  let m: Record<string, unknown> = {};
-  try {
-    const raw = item.change_payload;
-    if (raw) m = JSON.parse(raw) as Record<string, unknown>;
-  } catch {
-    /* keep empty */
-  }
-  const str = (v: unknown) => (typeof v === 'string' ? v : null);
-  return {
-    reason_text: item.reason_text ?? str(m.reason_text),
-    reason_code: item.reason_code ?? str(m.reason_code),
-    actor_id: item.actor_id ?? str(m.actor_id),
-    source: item.source ?? str(m.source),
-    correlation_id: item.correlation_id ?? str(m.correlation_id),
-    before_budget_period: item.before_budget_period ?? str(m.before_budget_period),
-    after_budget_period: item.after_budget_period ?? str(m.after_budget_period),
-    before_budget_reset_at: item.before_budget_reset_at ?? str(m.before_budget_reset_at),
-    after_budget_reset_at: item.after_budget_reset_at ?? str(m.after_budget_reset_at),
-  };
-}
-
-function isAuditObject(value: unknown): value is Record<string, unknown> {
-  return Boolean(value && typeof value === 'object' && !Array.isArray(value));
-}
-
-/** 与网关金额精度一致，用于判断 budget_max / budget_base 是否变化 */
-function budgetMoneySemanticallyEqual(
-  before: number | null | undefined,
-  after: number | null | undefined
-): boolean {
-  if (before == null && after == null) return true;
-  if (before == null || after == null) return false;
-  return before.toFixed(GATEWAY_MONEY_DECIMAL_PLACES) === after.toFixed(GATEWAY_MONEY_DECIMAL_PLACES);
-}
-
-function budgetResetAtSemanticallyEqual(
-  before: string | null | undefined,
-  after: string | null | undefined
-): boolean {
-  if ((before == null || before === '') && (after == null || after === '')) return true;
-  if (!before || !after) return false;
-  const tb = new Date(before).getTime();
-  const ta = new Date(after).getTime();
-  if (Number.isNaN(tb) || Number.isNaN(ta)) return before === after;
-  return tb === ta;
-}
-
-const budgetPlanHighlight = {
-  before: 'rounded px-0.5 bg-amber-50 text-amber-900',
-  after: 'rounded px-0.5 bg-sky-50 text-sky-900',
-} as const;
-
-function AuditChangedPair({
-  changed,
-  before,
-  after,
-}: {
-  changed: boolean;
-  before: string;
-  after: string;
-}) {
-  if (!changed) {
-    return <>{after}</>;
-  }
-  return (
-    <>
-      <span className={budgetPlanHighlight.before}>{before}</span>
-      <span className="text-gray-400"> → </span>
-      <span className={budgetPlanHighlight.after}>{after}</span>
-    </>
-  );
-}
-
-type AuditDiffRow = {
-  group: 'snapshot' | 'payload';
-  field: string;
-  before: string;
-  after: string;
-};
-
-function parseAuditJsonObject(raw: string | null | undefined): Record<string, unknown> | null {
-  const trimmed = raw?.trim();
-  if (!trimmed) return null;
-  try {
-    const parsed = JSON.parse(trimmed) as unknown;
-    if (isAuditObject(parsed)) return parsed;
-  } catch {
-    /* keep null */
-  }
-  return null;
-}
-
-function parseAuditChangedFields(raw: string | null | undefined): string[] | null {
-  const trimmed = raw?.trim();
-  if (!trimmed) return null;
-  try {
-    const parsed = JSON.parse(trimmed) as unknown;
-    if (!Array.isArray(parsed)) return null;
-    return parsed.filter((value): value is string => typeof value === 'string' && value.length > 0);
-  } catch {
-    return null;
-  }
-}
-
-function formatAuditDiffValue(value: unknown): string {
-  if (value == null || value === '') return '—';
-  if (typeof value === 'string') return value;
-  if (typeof value === 'number' || typeof value === 'boolean') return String(value);
-  try {
-    return JSON.stringify(value, null, 2);
-  } catch {
-    return String(value);
-  }
-}
-
-function pushAuditDiffRow(rows: AuditDiffRow[], group: AuditDiffRow['group'], field: string, before: unknown, after: unknown) {
-  const beforeText = formatAuditDiffValue(before);
-  const afterText = formatAuditDiffValue(after);
-  if (beforeText === afterText) return;
-  rows.push({ group, field, before: beforeText, after: afterText });
-}
-
-function pushNestedAuditDiffRows(
-  rows: AuditDiffRow[],
-  group: AuditDiffRow['group'],
-  field: string,
-  before: unknown,
-  after: unknown,
-  depth = 0
-) {
-  if ((isAuditObject(before) || isAuditObject(after)) && depth < 4) {
-    const beforeObject = isAuditObject(before) ? before : {};
-    const afterObject = isAuditObject(after) ? after : {};
-    const keys = Array.from(new Set([...Object.keys(beforeObject), ...Object.keys(afterObject)]));
-    keys.forEach((key) => {
-      pushNestedAuditDiffRows(
-        rows,
-        group,
-        `${field}.${key}`,
-        beforeObject[key],
-        afterObject[key],
-        depth + 1
-      );
-    });
-    return;
-  }
-  pushAuditDiffRow(rows, group, field, before, after);
-}
-
-function appendSnapshotDiffRows(rows: AuditDiffRow[], item: GatewayApiKeyBudgetAuditLog) {
-  const before = parseAuditJsonObject(item.before_user_snapshot ?? null);
-  const after = parseAuditJsonObject(item.after_user_snapshot ?? null);
-  if (!before && !after) return;
-
-  const fields = parseAuditChangedFields(item.changed_fields ?? null);
-  const keys =
-    fields && fields.length > 0
-      ? fields
-      : Array.from(new Set([...Object.keys(before ?? {}), ...Object.keys(after ?? {})])).filter((key) => key !== 'id');
-  const omitted = new Set<string>(OMIT_AUDIT_LOG_SNAPSHOT_FIELDS);
-  keys.forEach((key) => {
-    if (omitted.has(key)) return;
-    pushNestedAuditDiffRows(rows, 'snapshot', key, before?.[key], after?.[key]);
-  });
-}
-
-function appendPayloadDiffRows(rows: AuditDiffRow[], raw: string | null | undefined) {
-  const payload = parseAuditJsonObject(raw);
-  if (!payload) return;
-
-  const handled = new Set<string>();
-  const status = payload.status;
-  if (isAuditObject(status) && ('from' in status || 'to' in status)) {
-    pushAuditDiffRow(rows, 'payload', 'status', status.from, status.to);
-    handled.add('status');
-  }
-
-  const metadata = payload.metadata;
-  if (isAuditObject(metadata)) {
-    handled.add('metadata');
-    const changes = metadata.changes;
-    if (isAuditObject(changes)) {
-      Object.entries(changes).forEach(([key, value]) => {
-        if (isAuditObject(value) && ('from' in value || 'to' in value)) {
-          pushAuditDiffRow(rows, 'payload', `metadata.${key}`, value.from, value.to);
-        } else {
-          pushAuditDiffRow(rows, 'payload', `metadata.${key}`, '—', value);
-        }
-      });
-    } else if ('from' in metadata || 'to' in metadata) {
-      pushAuditDiffRow(rows, 'payload', 'metadata', metadata.from, metadata.to);
-      pushNestedAuditDiffRows(rows, 'payload', 'metadata', metadata.from, metadata.to);
-    } else {
-      Object.entries(metadata).forEach(([key, value]) => {
-        if (key === 'operation') return;
-        pushAuditDiffRow(rows, 'payload', `metadata.${key}`, '—', value);
-      });
-    }
-  }
-
-  Object.entries(payload).forEach(([key, value]) => {
-    if (handled.has(key)) return;
-    if (key.startsWith('before_') || key.startsWith('after_')) return;
-    if (key === 'metadata_patch_keys' || shouldOmitChangePayloadDisplayLine(`${key}:`)) return;
-    if (isAuditObject(value) && ('from' in value || 'to' in value)) {
-      pushAuditDiffRow(rows, 'payload', key, value.from, value.to);
-    }
-  });
-
-  Object.keys(payload)
-    .filter((key) => key.startsWith('before_'))
-    .forEach((beforeKey) => {
-      const suffix = beforeKey.slice('before_'.length);
-      const afterKey = `after_${suffix}`;
-      if (!(afterKey in payload)) return;
-      if (shouldOmitChangePayloadDisplayLine(beforeKey) || shouldOmitChangePayloadDisplayLine(afterKey)) return;
-      pushAuditDiffRow(rows, 'payload', suffix, payload[beforeKey], payload[afterKey]);
-    });
-}
-
-function auditDiffRows(item: GatewayApiKeyBudgetAuditLog): AuditDiffRow[] {
-  const rows: AuditDiffRow[] = [];
-  appendSnapshotDiffRows(rows, item);
-  appendPayloadDiffRows(rows, item.change_payload);
-  const seen = new Set<string>();
-  return rows.filter((row) => {
-    const key = `${row.field}:${row.before}:${row.after}`;
-    if (seen.has(key)) return false;
-    seen.add(key);
-    return true;
-  });
-}
-
-function isSerializedAuditObject(value: string): boolean {
-  return isAuditObject(parseAuditJsonObject(value));
-}
-
-function auditSummaryDiffRows(rows: AuditDiffRow[]): AuditDiffRow[] {
-  return rows.filter((row) => {
-    if (row.field === 'metadata') return false;
-    if (isSerializedAuditObject(row.before) || isSerializedAuditObject(row.after)) return false;
-    return true;
-  });
 }
 
 export default function GatewayAuditLogsPage() {
@@ -687,7 +356,6 @@ export default function GatewayAuditLogsPage() {
     });
     setPage(1);
   };
-  const detailRows = detailLog ? auditDiffRows(detailLog) : [];
 
   return (
     <div className="p-8">
@@ -925,247 +593,16 @@ export default function GatewayAuditLogsPage() {
                     </td>
                   </tr>
                 ) : (
-                  logs.map((item) => {
-                    const ex = auditDisplayExtras(item);
-                    const maxChanged = !budgetMoneySemanticallyEqual(
-                      item.before_budget_max,
-                      item.after_budget_max
-                    );
-                    const baseChanged = !budgetMoneySemanticallyEqual(
-                      item.before_budget_base,
-                      item.after_budget_base
-                    );
-                    const spentChanged = !budgetMoneySemanticallyEqual(item.before_spent, item.after_spent);
-                    const periodChanged =
-                      (ex.before_budget_period ?? '') !== (ex.after_budget_period ?? '');
-                    const resetChanged = !budgetResetAtSemanticallyEqual(
-                      ex.before_budget_reset_at,
-                      ex.after_budget_reset_at
-                    );
-                    const walletDiff = summarizeWalletSnapshotDiff(
-                      item.before_user_snapshot,
-                      item.after_user_snapshot
-                    );
-                    const reasonDisplay = auditReasonOneLine(ex.reason_code, ex.reason_text);
-                    const diffRows = auditDiffRows(item);
-                    const summaryDiffRows = auditSummaryDiffRows(diffRows);
-                    return (
+                  logs.map((item) => (
                     <tr key={item.id} className="hover:bg-gray-50">
-                      <td className="px-3 py-2 align-top">
-                        <div className="text-gray-700 whitespace-nowrap">{formatAuditTime(item.created_at, businessTimezone)}</div>
-                        <div
-                          className="mt-0.5 font-mono text-xs text-gray-600 whitespace-nowrap"
-                          title={item.request_log_id || undefined}
-                        >
-                          {t('labels.req')}: {item.request_log_id ? shortId(item.request_log_id) : '—'}
-                        </div>
-                        {ex.correlation_id ? (
-                          <div
-                            className="mt-0.5 font-mono text-xs text-gray-500 whitespace-nowrap"
-                            title={ex.correlation_id}
-                          >
-                            {t('labels.corr')}: {shortId(ex.correlation_id)}
-                          </div>
-                        ) : null}
-                      </td>
-                      <td className="px-3 py-2 align-top min-w-0 max-w-[15rem]">
-                        <div className="text-xs space-y-1.5 leading-snug">
-                          <div className="min-w-0">
-                            <span className="text-gray-500">{t('labels.type')}</span>
-                            <span className="text-sm font-medium text-gray-900" title={item.event_type}>
-                              {auditEnumLabel(t, 'eventTypes', item.event_type)}
-                            </span>
-                          </div>
-                          <div className="min-w-0 truncate text-[11px]" title={ex.source || undefined}>
-                            <span className="text-gray-500">{t('labels.from')}</span>
-                            <span className="text-violet-800">{auditEnumLabel(t, 'sourceChannels', ex.source)}</span>
-                          </div>
-                          <div className="min-w-0 line-clamp-3 text-gray-800" title={reasonDisplay.title || undefined}>
-                            <span className="text-gray-500">{t('labels.reason')}</span>
-                            <span className={reasonDisplay.isMono ? 'font-mono text-[11px] text-gray-900' : 'text-[11px]'}>
-                              {reasonDisplay.line}
-                            </span>
-                          </div>
-                        </div>
-                      </td>
-                      <td className="px-3 py-2 align-top min-w-0 max-w-[12rem]">
-                        <div className="text-xs space-y-1.5 leading-snug">
-                          <div>
-                            <span className="text-gray-500">{t('labels.kind')}</span>
-                            <span className="text-sm text-gray-900" title={item.actor_type}>
-                              {auditEnumLabel(t, 'actorTypes', item.actor_type)}
-                            </span>
-                          </div>
-                          <div className="min-w-0">
-                            <span className="text-gray-500">{t('labels.principal')}</span>
-                            {ex.actor_id ? (
-                              (() => {
-                                const { kind, identifier } = parseAuditActorId(ex.actor_id);
-                                return (
-                                  <span className="inline-flex min-w-0 flex-wrap items-baseline gap-1" title={ex.actor_id}>
-                                    {kind ? (
-                                      <span className="rounded bg-gray-100 px-1 text-[10px] text-gray-700" title={kind}>
-                                        {auditEnumLabel(t, 'actorKinds', kind)}
-                                      </span>
-                                    ) : null}
-                                    <span className="font-mono text-[11px] text-gray-700 break-all">
-                                      {shortId(identifier)}
-                                    </span>
-                                  </span>
-                                );
-                              })()
-                            ) : (
-                              <span className="text-gray-400">—</span>
-                            )}
-                          </div>
-                        </div>
-                      </td>
-                      <td className="px-3 py-2 align-top min-w-0 max-w-[18rem]">
-                        <div className="text-sm text-gray-900 truncate leading-snug" title={item.user_email || ''}>
-                          {item.user_email || '—'}
-                        </div>
-                        {item.user_id ? (
-                          <div className="mt-0.5 flex items-baseline gap-1 min-w-0 font-mono text-xs">
-                            <span className="shrink-0 text-gray-600">{t('labels.user')}</span>
-                            <Link
-                              href={`/gateway/users/${encodeURIComponent(item.user_id)}`}
-                              className="min-w-0 truncate text-blue-600 hover:underline"
-                              title={item.user_id}
-                            >
-                              {shortId(item.user_id)}
-                            </Link>
-                          </div>
-                        ) : (
-                          <div className="mt-0.5 font-mono text-xs text-gray-400 truncate" title={t('userRemovedTitle')}>
-                            {t('labels.user')} —
-                          </div>
-                        )}
-                        <div className="mt-0.5 font-mono text-xs text-gray-500 truncate leading-snug" title={item.api_key_id ?? ''}>
-                          {t('labels.key')}{item.api_key_id ? shortId(item.api_key_id) : '—'}
-                        </div>
-                      </td>
-                      <td className="px-3 py-2 align-top text-xs text-gray-600 min-w-[14rem]">
-                        <div className="space-y-1 leading-snug">
-                          <div>
-                            <span className="font-medium text-gray-700">{t('labels.spent')}</span>{' '}
-                            {spentChanged ? (
-                              <span className="inline-block align-top">
-                                <AuditChangedPair
-                                  changed
-                                  before={formatGatewayMoneyCode(item.before_spent, billingCurrency, GATEWAY_MONEY_DECIMAL_PLACES)}
-                                  after={formatGatewayMoneyCode(item.after_spent, billingCurrency, GATEWAY_MONEY_DECIMAL_PLACES)}
-                                />
-                                <div
-                                  className={
-                                    item.delta_spent > 0
-                                      ? 'text-red-600'
-                                      : item.delta_spent < 0
-                                        ? 'text-green-600'
-                                        : 'text-gray-500'
-                                  }
-                                >
-                                  {formatSignedMoney(item.delta_spent, billingCurrency)}
-                                </div>
-                              </span>
-                            ) : (
-                              formatGatewayMoneyCode(item.after_spent, billingCurrency, GATEWAY_MONEY_DECIMAL_PLACES)
-                            )}
-                          </div>
-                          <div>
-                            <span className="font-medium text-gray-700">{t('labels.max')}</span>{' '}
-                            <AuditChangedPair
-                              changed={maxChanged}
-                              before={formatBudgetMax(item.before_budget_max, billingCurrency)}
-                              after={formatBudgetMax(item.after_budget_max, billingCurrency)}
-                            />
-                          </div>
-                          <div>
-                            <span className="font-medium text-gray-700">{t('labels.base')}</span>{' '}
-                            <AuditChangedPair
-                              changed={baseChanged}
-                              before={formatGatewayMoneyCode(item.before_budget_base, billingCurrency, GATEWAY_MONEY_DECIMAL_PLACES)}
-                              after={formatGatewayMoneyCode(item.after_budget_base, billingCurrency, GATEWAY_MONEY_DECIMAL_PLACES)}
-                            />
-                          </div>
-                        </div>
-                      </td>
-                      <td className="px-3 py-2 align-top text-xs text-gray-600 min-w-[12rem]">
-                        <div className="space-y-1 leading-snug">
-                          <div>
-                            <span className="font-medium text-gray-700">{t('labels.period')}</span>{' '}
-                            <AuditChangedPair
-                              changed={periodChanged}
-                              before={ex.before_budget_period ?? '—'}
-                              after={ex.after_budget_period ?? '—'}
-                            />
-                          </div>
-                          <div>
-                            <span className="font-medium text-gray-700">{t('labels.resetAt')}</span>{' '}
-                            <span className="whitespace-nowrap">
-                              <AuditChangedPair
-                                changed={resetChanged}
-                                before={formatAuditTime(ex.before_budget_reset_at, businessTimezone)}
-                                after={formatAuditTime(ex.after_budget_reset_at, businessTimezone)}
-                              />
-                            </span>
-                          </div>
-                        </div>
-                      </td>
-                      <td className="px-3 py-2 align-top text-xs text-gray-600 min-w-[14rem]">
-                        <AuditWalletPlanBlock
-                          diff={walletDiff}
-                          currency={billingCurrency}
-                          labels={{
-                            granted: t('labels.granted'),
-                            spent: t('labels.spent'),
-                            remaining: t('labels.remaining'),
-                          }}
-                        />
-                      </td>
-                      <td className="px-3 py-2 text-gray-600 min-w-[14rem] max-w-lg align-top">
-                        <div className="space-y-1 text-xs leading-snug">
-                          {summaryDiffRows.length > 0 ? (
-                            <>
-                              {summaryDiffRows.slice(0, 3).map((row, index) => (
-                                <div key={`${item.id}-diff-${index}`} className="grid grid-cols-[minmax(5rem,8rem)_1fr] gap-x-2">
-                                  <span className="truncate font-mono text-gray-700" title={row.field}>
-                                    {row.field}
-                                  </span>
-                                  <span className="min-w-0 truncate" title={`${row.before} → ${row.after}`}>
-                                    <span className="text-amber-700">{row.before}</span>
-                                    <span className="px-1 text-gray-400">→</span>
-                                    <span className="text-sky-700">{row.after}</span>
-                                  </span>
-                                </div>
-                              ))}
-                              <button
-                                type="button"
-                                onClick={() => setDetailLog(item)}
-                                className="mt-1 text-xs text-blue-600 hover:underline"
-                              >
-                                {t('labels.viewChangeDetail')}
-                                {summaryDiffRows.length > 3 ? ` · ${t('labels.more', { count: summaryDiffRows.length - 3 })}` : ''}
-                              </button>
-                            </>
-                          ) : (
-                            <>
-                              <span className="text-gray-400">—</span>
-                              {diffRows.length > 0 ? (
-                                <button
-                                  type="button"
-                                  onClick={() => setDetailLog(item)}
-                                  className="ml-2 text-xs text-blue-600 hover:underline"
-                                >
-                                  {t('labels.viewChangeDetail')}
-                                </button>
-                              ) : null}
-                            </>
-                          )}
-                        </div>
-                      </td>
+                      <AuditLogSharedCells
+                        item={item}
+                        currency={billingCurrency}
+                        timezone={businessTimezone}
+                        onViewDetail={setDetailLog}
+                      />
                     </tr>
-                    );
-                  })
+                  ))
                 )}
               </tbody>
             </table>
@@ -1174,65 +611,11 @@ export default function GatewayAuditLogsPage() {
       )}
 
       {detailLog ? (
-        <div
-          className="fixed inset-0 z-50 flex items-center justify-center bg-black/50 p-4"
-          role="dialog"
-          aria-modal="true"
-          aria-labelledby="audit-change-detail-title"
-        >
-          <div className="flex max-h-[85vh] w-full max-w-5xl flex-col overflow-hidden rounded-lg bg-white shadow-xl">
-            <div className="flex items-start justify-between gap-4 border-b border-gray-200 px-5 py-4">
-              <div className="min-w-0">
-                <h2 id="audit-change-detail-title" className="text-lg font-semibold text-gray-900">
-                  {t('labels.changeDetailTitle')}
-                </h2>
-                <div className="mt-1 flex flex-wrap gap-x-3 gap-y-1 text-xs text-gray-500">
-                  <span title={detailLog.event_type}>{auditEnumLabel(t, 'eventTypes', detailLog.event_type)}</span>
-                  <span>{formatAuditTime(detailLog.created_at, businessTimezone)}</span>
-                  <span className="truncate">{detailLog.user_email ?? '—'}</span>
-                </div>
-              </div>
-              <button
-                type="button"
-                onClick={() => setDetailLog(null)}
-                className="rounded-md border border-gray-300 px-3 py-1.5 text-sm text-gray-700 hover:bg-gray-50"
-              >
-                {tCommon('close')}
-              </button>
-            </div>
-
-            <div className="overflow-y-auto p-5">
-              {detailRows.length > 0 ? (
-                <div className="space-y-3">
-                  {detailRows.map((row, index) => (
-                    <div key={`${row.group}-${row.field}-${index}`} className="rounded-lg border border-gray-200 bg-white p-3">
-                      <div className="mb-2 flex flex-wrap items-center gap-2">
-                        <span className="rounded bg-gray-100 px-2 py-0.5 text-xs text-gray-600">
-                          {row.group === 'snapshot' ? t('labels.userSnapshot') : t('labels.extraJson')}
-                        </span>
-                        <span className="min-w-0 break-all font-mono text-sm font-medium text-gray-900">
-                          {row.field}
-                        </span>
-                      </div>
-                      <div className="grid gap-3 md:grid-cols-2">
-                        <div className="min-w-0">
-                          <div className="mb-1 text-xs font-medium uppercase text-gray-500">{t('labels.originalValue')}</div>
-                          <pre className="max-w-full whitespace-pre-wrap break-all rounded bg-amber-50 px-3 py-2 font-mono text-xs leading-relaxed text-amber-900">{row.before}</pre>
-                        </div>
-                        <div className="min-w-0">
-                          <div className="mb-1 text-xs font-medium uppercase text-gray-500">{t('labels.changedValue')}</div>
-                          <pre className="max-w-full whitespace-pre-wrap break-all rounded bg-sky-50 px-3 py-2 font-mono text-xs leading-relaxed text-sky-900">{row.after}</pre>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              ) : (
-                <div className="py-8 text-center text-sm text-gray-500">{t('labels.noChangeDetail')}</div>
-              )}
-            </div>
-          </div>
-        </div>
+        <AuditChangeDetailModal
+          item={detailLog}
+          timezone={businessTimezone}
+          onClose={() => setDetailLog(null)}
+        />
       ) : null}
 
       {totalPages > 1 && !isLoading && (
